requirements.txt
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6-sip==13.10.2
opencv-python>=4.10.0.84
numpy>=2.1
requests>=2.31
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2.7
Pillow>=10.0
psutil>=5.9
